move hand evaluator to deck.js
